+++ frontend/index.html
@@ -2,46 +2,46 @@
 <html>
 
 <head>
-    <meta charset="utf-8" />
-      <title>SLICER e-Shop</title>
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;800&display=swap" rel="stylesheet">
-    <link href="style.css" rel="stylesheet">
+  <meta charset="utf-8" />
+  <title>SLICER e-Shop</title>
+  <link rel="preconnect" href="https://fonts.gstatic.com">
+  <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;800&display=swap" rel="stylesheet">
+  <link href="style.css" rel="stylesheet">
 </head>
 
 <body class="align-center column">
-    <header class="align-center header">SLICER</header>
+  <header class="align-center header">SLICER</header>
 
-      <main class="align-center column main" id="content">
-          <div class="container align-center">
-            <div class="align-center column modal" id="modal">
-              <button class="modal__button" id="close">
-                <i class="fas fa-times-circle"></i>
-              </button>
-                <div class="card" id="details"></div>
-              <button class="modal__button" id="add">
-                <i class="fas fa-cart-plus"></i>
-              </button>
+  <main class="align-center column main" id="content">
+    <div class="container align-center">
+      <div class="align-center column modal" id="modal">
+        <button class="modal__button" id="close">
+          <i class="fas fa-times-circle"></i>
+        </button>
+        <div class="card" id="details"></div>
+        <button class="modal__button" id="add">
+          <i class="fas fa-cart-plus"></i>
+        </button>
 
-              <div class="modal__footer"></div>
-            </div>
+        <div class="modal__footer"></div>
+      </div>
 
-            <ul class="align-center column" class="list" id="list"></ul>
-          </div>
+      <ul class="align-center column" class="list" id="list"></ul>
+    </div>
 
-          <div class="align-center pagination" id="pagination">
-            <button class="fas fa-chevron-left pagination__button" id="prev"></button>
-              <div class="align-center pagination__number" id="num">1</div>
-            <button class="fas fa-chevron-right pagination__button" id="next"></button>
-          </div>
-      </main>
+    <div class="align-center pagination" id="pagination">
+      <button class="fas fa-chevron-left pagination__button" id="prev"></button>
+      <div class="align-center pagination__number" id="num">1</div>
+      <button class="fas fa-chevron-right pagination__button" id="next"></button>
+    </div>
+  </main>
 
-    <footer class="align-center footer">
-        <small>2021 SLICER &copy;</small>
-    </footer>
+  <footer class="align-center footer">
+    <small>2021 SLICER &copy;</small>
+  </footer>
 
-    <script src="https://kit.fontawesome.com/e9e0221c2b.js" crossorigin="anonymous"></script>
-    <script src="./prod/script.js" type="module"></script>
+  <script src="https://kit.fontawesome.com/e9e0221c2b.js" crossorigin="anonymous"></script>
+  <script src="./prod/controller.js" type="module"></script>
 </body>
 
 </html>
