+++ package.json
@@ -20,7 +20,8 @@
   "homepage": "https://github.com/DmitryKorolDevPro/MentorshipProgramRemoteProject#readme",
   "dependencies": {
     "cors": "2.8.5",
-    "express": "4.17.1"
+    "express": "4.17.1",
+    "lint": "0.7.0"
   },
   "devDependencies": {
     "@mate-academy/eslint-config": "0.0.16",
@@ -31,6 +32,7 @@
     "@typescript-eslint/parser": "4.22.1",
     "eslint": "7.2.0",
     "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-node": "11.1.0",
     "live-server": "1.2.1",
     "nodemon": "2.0.7",
     "typescript": "4.2.4"
